Rename Interfaces folder and update namespaces
