+++ UserManagementSystem.Api/Controllers/AuthController.cs
@@ -6,7 +6,7 @@ using System.Security.Cryptography;
 using System.Text;
 using UserManagementSystem.Api.Models;
 using UserManagementSystem.Application.DTOs.Responses;
-using UserManagementSystem.Application.Interfcaces;
+using UserManagementSystem.Application.Interfaces;
 using UserManagementSystem.Domain.Entities;
 
 namespace UserManagementSystem.Api.Controllers
