+++ UserManagementSystem.Application/Interfaces/IUserProfileService.cs
